Scala 2.10.4 final release.
Add announcement with release notes, the download page for 2.10.4, update
index of download section, update link to latest scaladoc, update the
`_config.yml` file. All those steps extracted from Scala 2.11.0-RC3 release
commits.

Note, that download/index.md is not auto generated despite the comment in
that file suggesting otherwise. When updating that file, I corrected the
size of tgz file (it was 20MB where it should be 29MB for both 2.10.3
and 2.10.4).