Replace inline display of community tickets with a JIRA search
The inline display has been broken for a while since a JIRA upgrade.

While I'm not an expert in Javascript, I believe that we can't
get this working until https://jira.atlassian.com/browse/JRA-30371
is fixed. Well, we could with a server that proxied the JIRA
REST API and manipulated headers, but that's too much work
for this page.

Instead, I've linked to JIRA searches to show open and fixed
community tickets, and added some explanatory comments for would-be
contributors.

Fixes #209