Fix the front-page title
The current title is pretty strange; it really doesn't mean anything to me, someone who has been using Scala for a long time, and would mean even less to a newcomer. We should dispense with the jargon and say what people would be interested in hearing, in plain language that any programmer can understand.