Jekyll 3.9.0 upgrade, also kramdown upgrade
this is comparable to scala/docs.scala-lang#1756.  the motivation is
the same: we got a security alert about kramdown.

but it went a bit differently in this repo. if I did `bundle update
jekyll` that went all the way to some 4.x version. I don't know why.
but it seems safer to do a smaller upgrade first. `bundle update
--minor jekyll` got us to 3.9.0, but then kramdown didn't get updated,
so I followed it with `bundle update kramdown` (after adding
`kramdown-parser-gfm` to `Gemfile`, as in the other PR)

since I didn't do a full `bundle update`, I didn't need to revert
any version bump of jekyll-redirect-from

so.... it's not great that this doesn't bring the two repos exactly
in sync, but oh well. I'm trying to resolve the security issue without
putting excess effort into it.

I wouldn't oppose followup PR(s) that reduce the deltas between the
repos.