Install page - Remove Scastie references.
Scastie is not a way for _installing_ or downloading Scala into a
local machine, which is what the title of the page says.
There are links to Scastie directly from main page.

Also, when I tried the link to Scalafiddle it seem not to be working.
If it is another playground like Scastie,
should it be linked from the scala-lang page?